ya el carrito puede actualizar la cantidad de un producto

+++ src/Controller/ProductosController.php
@@ -230,15 +230,32 @@ class ProductosController extends AppController
         if($this->request->is('post')) {
             $DatosBoton = TableRegistry::get('carrito_compras')->patchEntity($DatosBoton, $this->request->data);
             $LlavePrimaria = array($DatosBoton['idPersona'],$DatosBoton['idProducto']);
-            $TuplaBorrar = TableRegistry::get('carrito_compras')-> get ($LlavePrimaria);
+            
+            if (isset($this->request->data['BotonActualizarCarrito'])) {
+                $TuplaActualizar = TableRegistry::get('carrito_compras')-> get ($LlavePrimaria);
+                $TuplaActualizar->set('cantidad', $DatosBoton['cantidad']);
+                
+                 if(TableRegistry::get('carrito_compras')->save($TuplaActualizar)) {
+                    $this->Flash->success('La cantidad del producto ha sido actualizada.');
+                    return $this->redirect(['action' => '../carrito/'.$codigo]);
+                }
+                else {
+                    $this->Flash->error('El producto no se ha actualizado debido a un error.');
+                }
+                
+            } else if (isset($this->request->data['BotonBorrar'])) {
+                $TuplaBorrar = TableRegistry::get('carrito_compras')-> get ($LlavePrimaria);
 
-            if(TableRegistry::get('carrito_compras')->delete($TuplaBorrar)) {
-                $this->Flash->success('Producto borrado del carrito de compras.');
-                return $this->redirect(['action' => '../carrito/'.$codigo]);
+                if(TableRegistry::get('carrito_compras')->delete($TuplaBorrar)) {
+                    $this->Flash->success('Producto borrado del carrito de compras.');
+                    return $this->redirect(['action' => '../carrito/'.$codigo]);
+                }
+                else {
+                    $this->Flash->error('El producto no se ha borrado debido a un error.');
+                }
+                
             }
-            else {
-                $this->Flash->error('El producto no se ha borrado debido a un error.');
-            }
+            
         }
 
         $this->set(compact('DatosBoton'));
